Move conversions to different headers to minimise deps brought in
