+++ gazebo_ros/include/gazebo_ros/conversions/sensor_msgs.hpp
@@ -12,21 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GAZEBO_ROS__CONVERSIONS_HPP_
-#define GAZEBO_ROS__CONVERSIONS_HPP_
+#ifndef GAZEBO_ROS__CONVERSIONS__SENSOR_MSGS_HPP_
+#define GAZEBO_ROS__CONVERSIONS__SENSOR_MSGS_HPP_
 
 #include <math.h>
+#include "gazebo_ros/conversions/builtin_interfaces.hpp"
+#include "gazebo_ros/conversions/generic.hpp"
 
-#include <builtin_interfaces/msg/time.hpp>
-#include <gazebo/common/Time.hh>
 #include <gazebo/msgs/laserscan_stamped.pb.h>
 #include <geometry_msgs/msg/point32.hpp>
-#include <geometry_msgs/msg/quaternion.hpp>
-#include <geometry_msgs/msg/vector3.hpp>
-#include <ignition/math/Quaternion.hh>
-#include <ignition/math/Vector3.hh>
 #include <rclcpp/logging.hpp>
-#include <rclcpp/time.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
 #include <sensor_msgs/msg/point_cloud.hpp>
 #include <sensor_msgs/msg/laser_scan.hpp>
@@ -37,179 +32,15 @@
 #include <sstream>  // NOLINT
 #include <sensor_msgs/point_cloud2_iterator.hpp>  // NOLINT
 
-#include <string>
 #include <algorithm>
 #include <limits>
 
 namespace gazebo_ros
 {
+namespace conversions
+{
 /// \brief Common logger for all conversions
 static rclcpp::Logger conversions_logger = rclcpp::get_logger("gazebo_ros_conversions");
-
-/// Generic conversion from a ROS geometry vector message to another type.
-/// \param[in] in Input message.
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const geometry_msgs::msg::Vector3 & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from a ROS vector message to an Ignition Math vector.
-/// \param[in] msg ROS message to convert.
-/// \return An Ignition Math vector.
-template<>
-ignition::math::Vector3d Convert(const geometry_msgs::msg::Vector3 & msg)
-{
-  ignition::math::Vector3d vec;
-  vec.X(msg.x);
-  vec.Y(msg.y);
-  vec.Z(msg.z);
-  return vec;
-}
-
-/// Generic conversion from a ROS geometry point message to another type.
-/// \param[in] in Input message.
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const geometry_msgs::msg::Point32 & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from a ROS point message to an Ignition Math vector.
-/// \param[in] in ROS message to convert.
-/// \return An Ignition Math vector.
-template<>
-ignition::math::Vector3d Convert(const geometry_msgs::msg::Point32 & in)
-{
-  ignition::math::Vector3d vec;
-  vec.X(in.x);
-  vec.Y(in.y);
-  vec.Z(in.z);
-  return vec;
-}
-
-/// Generic conversion from an Ignition Math vector to another type.
-/// \param[in] in Input vector.
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const ignition::math::Vector3d & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from an Ignition Math vector to a ROS message.
-/// \param[in] vec Ignition vector to convert.
-/// \return ROS geometry vector message
-template<>
-geometry_msgs::msg::Vector3 Convert(const ignition::math::Vector3d & vec)
-{
-  geometry_msgs::msg::Vector3 msg;
-  msg.x = vec.X();
-  msg.y = vec.Y();
-  msg.z = vec.Z();
-  return msg;
-}
-
-/// Generic conversion from an Ignition Math quaternion to another type.
-/// \param[in] in Input quaternion
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const ignition::math::Quaterniond & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from an Ignition Math Quaternion to a ROS message.
-/// \param[in] in Ignition Quaternion to convert.
-/// \return ROS geometry quaternion message
-template<>
-geometry_msgs::msg::Quaternion Convert(const ignition::math::Quaterniond & in)
-{
-  geometry_msgs::msg::Quaternion msg;
-  msg.x = in.X();
-  msg.y = in.Y();
-  msg.z = in.Z();
-  msg.w = in.W();
-  return msg;
-}
-
-/// Generic conversion from a ROS Quaternion message to another type
-/// \param[in] in Input quaternion
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const geometry_msgs::msg::Quaternion & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from a ROS quaternion message to ignition quaternion
-/// \param[in] in Input quaternion message
-/// \return Ignition math quaternion with same values as the input message
-template<>
-ignition::math::Quaterniond Convert(const geometry_msgs::msg::Quaternion & in)
-{
-  return ignition::math::Quaterniond(in.w, in.x, in.y, in.z);
-}
-
-/// Generic conversion from an Gazebo Time object to another type.
-/// \param[in] in Input time;
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const gazebo::common::Time & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from an Gazebo Time to a RCLCPP Time.
-/// \param[in] in Gazebo Time to convert.
-/// \return A rclcpp::Time object with the same value as in
-template<>
-rclcpp::Time Convert(const gazebo::common::Time & in)
-{
-  return rclcpp::Time(in.sec, in.nsec, rcl_clock_type_t::RCL_ROS_TIME);
-}
-
-/// \brief Specialized conversion from an Gazebo Time to a ROS Time message.
-/// \param[in] in Gazebo Time to convert.
-/// \return A ROS Time message with the same value as in
-template<>
-builtin_interfaces::msg::Time Convert(const gazebo::common::Time & in)
-{
-  builtin_interfaces::msg::Time time;
-  time.sec = in.sec;
-  time.nanosec = in.nsec;
-  return time;
-}
-
-/// Generic conversion from an Gazebo Time message to another type.
-/// \param[in] in Input time
-/// \return Conversion result
-/// \tparam OUT Output type
-template<class OUT>
-OUT Convert(const gazebo::msgs::Time & in)
-{
-  return OUT();
-}
-
-/// \brief Specialized conversion from an Gazebo Time message to a ROS Time message.
-/// \param[in] in Gazebo Time message to convert.
-/// \return A ROS Time message with the same value as in
-template<>
-builtin_interfaces::msg::Time Convert(const gazebo::msgs::Time & in)
-{
-  builtin_interfaces::msg::Time time;
-  time.sec = in.sec();
-  time.nanosec = in.nsec();
-  return time;
-}
 
 /// Generic conversion from an Gazebo Laser Scan message to another type.
 /// \param[in] in Input message;
@@ -485,5 +316,6 @@ sensor_msgs::msg::Range Convert(const gazebo::msgs::LaserScanStamped & in, doubl
   return range_msg;
 }
 
+}  // namespace conversions
 }  // namespace gazebo_ros
-#endif  // GAZEBO_ROS__CONVERSIONS_HPP_
+#endif  // GAZEBO_ROS__CONVERSIONS__SENSOR_MSGS_HPP_
